api/poc-api/requirements/base.txt
Django==2.2.7
psycopg2-binary==2.8.4
django-filter==2.2.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
django-rest-swagger==2.2.0
Markdown==3.1.1
pytz==2019.3
sqlparse==0.3.0
django-environ==0.4.5
ipython==7.9.0
django-unixdatetimefield==1.0.3
PyYAML==5.1.2
django-allauth==0.41.0
Pillow

# Tests
factory-boy==2.12.0
django-test-plus==1.4.0
coverage==4.5.4